scripts/pyproject.toml
[project]
name = "build-thumbbar-icons"
version = "0.1.0"
description = "Build Windows .ico files for taskbar thumbnail toolbar icons from SVGs"
requires-python = ">=3.9"
dependencies = [
    "pillow>=10.0.0",
    "resvg-py>=0.2.6",
]

[project.scripts]
build-thumbbar = "build_thumbbar_icons:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
